refactor(fe): extract ProfileSection
